requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
boto3>=1.20.0
torch>=1.10.0
numpy>=1.21.0
monai==1.4.0
